+++ rust/wgpu-mc-demo/src/main.rs
@@ -319,7 +319,7 @@ fn begin_rendering(event_loop: EventLoop<()>, window: Window, wm: WmRenderer) {
                                 },
                                 PartTransform {
                                     x: 0.0,
-                                    y: ((spin / 20.0).sin() * 0.5) as f32,
+                                    y: ((spin / 20.0).sin() * 0.5),
                                     // y: 0.0,
                                     z: 0.0,
                                     pivot_x: 0.5,
@@ -334,7 +334,7 @@ fn begin_rendering(event_loop: EventLoop<()>, window: Window, wm: WmRenderer) {
                                 },
                                 PartTransform {
                                     x: 0.0,
-                                    y: ((spin / 20.0).sin() * 0.5) as f32,
+                                    y: ((spin / 20.0).sin() * 0.5),
                                     // y: 0.0,
                                     z: 0.0,
                                     pivot_x: 0.6,
@@ -349,7 +349,7 @@ fn begin_rendering(event_loop: EventLoop<()>, window: Window, wm: WmRenderer) {
                                 },
                                 PartTransform {
                                     x: 0.0,
-                                    y: ((spin / 20.0).sin() * 0.5) as f32,
+                                    y: ((spin / 20.0).sin() * 0.5),
                                     // y: 0.0,
                                     z: 0.0,
                                     pivot_x: 0.6,
